Chat Room Messages CRUD

+++ base/views.py
@@ -1,17 +1,27 @@
 from django.contrib import messages
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-from .models import Room, Topic
+from django.contrib.auth.forms import UserCreationForm
+
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 # Create your views here.
 
+
 def loginPage(request):
 
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username =request.POST.get('username')
+        username =request.POST.get('username').lower()
         password =request.POST.get('password')
 
         try:
@@ -28,16 +38,40 @@ def loginPage(request):
             messages.error(request,"Username or password does not exist")
 
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
+
+
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration.')
+
+
+    return render(request, 'base/login_register.html', {'form':form})
+
+
+
+
 def home(request):
 
-    q= request.GET.get('q') if request.GET.get('q') != None else ''
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(Q(topic__name__icontains=q) | 
                                 Q(name__icontains=q) | 
                                 Q(description__icontains=q)
@@ -49,12 +83,29 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics , 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
+
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
     return render(request, 'base/room.html', context)
 
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -66,9 +117,16 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to edit this room')
+
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -79,9 +137,29 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html',context)
 
+
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not authorized to edit this room')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != Message.user:
+        return HttpResponse('You are not authorized to edit this room')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj':message})
